requirements.txt
newspaper3k==0.2.8
sanic==19.12.2
aiohttp==3.6.2
